LayoutTests/imported/w3c/web-platform-tests/resize-observer/observe-expected.txt

CONSOLE MESSAGE: line 50: ReferenceError: Can't find variable: ResizeObserver
ResizeObserver tests

t1
t2
 

FAIL ResizeObserver implemented assert_own_property: expected property "ResizeObserver" missing
